Move database operations out of command
A new StackRepository is introduced which is responsible
for the database operations. The URL records are yielded to
preserve memory.

The output of the command was improved, now the total number
of processed URLs along with skipped and erroneous URLs is given.

The SearchEngineService was renamed to SearchEngineNotifier
to make its purpose clearer.